requirements.txt
attrs==22.1.0
certifi==2022.9.14
charset-normalizer==2.1.1
Django==2.2.28
djangorestframework==3.13.1
idna==3.4
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
pytest-django==4.5.2
pytest-html==3.1.1
pytest-metadata==2.0.2
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.12
